fix the transaction provider by calculating latest N transaction at client-side instead of server side, added new Activity page to show all activity

+++ duitwise_app/lib/modules/financial_tracking/view/mybudget_page.dart
@@ -247,7 +247,7 @@ class _AddTransactionPopupState extends ConsumerState<AddTransactionPopup> {
           notes: notes,
           category: selectedCategory,
           amount: amount,
-          createdAt: DateTime.now().millisecondsSinceEpoch.toDouble(), // use seconds instead of millisecond
+          createdAt: DateTime.now().millisecondsSinceEpoch.toInt(), // use seconds instead of millisecond
         ),
       );
